refactor: move IconLink to its own file (#3170)
Fix #3170
